Reorganize `utils.testing` code
